+++ discord/_dataclass.py
@@ -1,5 +1,5 @@
 from .flags import PermissionFlags
-from .objects.components._base import Component, LabelChildComponent
+from .objects.components._base import Component, ContainerChildComponent, LabelChildComponent
 from .snowflake import Snowflake
 from .utils import ISO8601Timestamp, Match, MISSING, Nullable, Optional
 from annotationlib import get_annotations
@@ -33,7 +33,7 @@ def dataclass[T](cls: T) -> T:
       if annotation.__origin__ is Match: return self.__parse_match(annotation.__args__[1:], annotation.__args__[0], value, data = data)
       if annotation.__origin__ is Nullable: return self.__parse_nullable(annotation.__args__[0], value, data = data)
       if annotation.__origin__ is Optional: return self.__parse_optional(annotation.__args__[0], value, data = data)
-    if annotation in (Component, LabelChildComponent): return annotation(**value)
+    if annotation in (Component, ContainerChildComponent, LabelChildComponent): return annotation(**value)
     if is_dataclass(annotation): return self.__parse_dataclass(annotation, value, data = data)
     if annotation is Snowflake and value not in (MISSING, None, Ellipsis): return Snowflake(value)
     if issubclass(annotation, Enum) and value not in (MISSING, None, Ellipsis):